Refactoring metric types & deserializations

+++ WiseOldManConnector/Models/API/Responses/WOMDelta.cs
@@ -2,8 +2,8 @@
 
 namespace WiseOldManConnector.Models.API.Responses {
     internal class WOMDelta {
-        public int Start { get; set; }
-        public int End { get; set; }
+        public double Start { get; set; }
+        public double End { get; set; }
 
         [JsonProperty("gained")]
         public double Gained { get; set; }
